pyproject.toml
[project]
name = "airi-research"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "facenet-pytorch>=2.6.0",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.7",
    "mlflow>=2.22.0",
    "moviepy>=2.2.1",
    "numpy<2",
    "pandas>=2.2.3",
    "pillow>=10.2.0",
    "pip>=25.1.1",
    "setuptools>=80.9.0",
    "torchinfo>=1.8.0",
    "torchvideo>=0.0.0",
    "ydata-profiling>=4.16.1",
]
